+++ openspec/changes/archive/2026-05-23-a29-split-webhook-dispatcher/design.md
